fix: drop zx for bin
