+++ package.json
@@ -12,7 +12,7 @@
     "test": "jest",
     "test:coverage": "jest --coverage",
     "build": "yarn t && bob build",
-    "gen": "zx bin/theme-gen.mjs bin/theme.ts --output bin/ret.d.ts",
+    "gen": "node bin/theme-gen.js bin/theme.ts bin/ret.d.ts",
     "prepare": "husky"
   },
   "repository": {
@@ -26,12 +26,15 @@
     "@commitlint/config-conventional": "^19.1.0",
     "@mj-studio/eslint-config-react": "^1.0.4",
     "@testing-library/react-hooks": "^8.0.1",
+    "@types/fs-extra": "^11",
     "@types/invariant": "^2.2.37",
     "@types/jest": "^29.5.12",
     "@types/react": "^18.2.65",
     "@types/react-test-renderer": "^18",
     "cz-conventional-changelog": "^3.3.0",
     "eslint": "^8.2.0",
+    "execa": "^8.0.1",
+    "fs-extra": "^11.2.0",
     "husky": "^9.0.11",
     "jest": "^29.7.0",
     "metro-react-native-babel-preset": "^0.77.0",
@@ -42,8 +45,7 @@
     "react-test-renderer": "^18.2.0",
     "semantic-release": "^23.0.2",
     "ts-jest": "^29.1.2",
-    "typescript": "5.3.3",
-    "zx": "^7.2.3"
+    "typescript": "5.3.3"
   },
   "peerDependencies": {
     "react": "*",
@@ -61,7 +63,7 @@
     "@mj-studio/js-util": "^1.0.29",
     "invariant": "^2.2.4"
   },
-  "bin": "./bin/theme-gen.mjs",
+  "bin": "./bin/theme-gen.js",
   "react-native": "src/index.ts",
   "source": "src/index.ts",
   "module": "lib/module/index.js",
